fix(vsprintf): give the length modifiers a 64-bit path that consumes the whole argument (#332) (#358)
The modifiers were parsed but every argument was read through a
4-byte va_arg, so %lld still left the cursor misaligned for the rest
of the call, and the z paths truncated through a stray (short) cast.
The number paths now carry unsigned long long end to end, ll reads
eight bytes, and z reads the full size_t/ssize_t.

64-bit division cannot use the compiler's helpers in a -nostdlib
libc, so the digit conversion divides with shifts and subtracts,
with a native division fast path for values that fit in 32 bits.

The four call sites that narrowed a uint64_t to printf it
(scheduler_feedback, /proc/uptime, timer debug, the ATA sectors_48
dump) now print the whole value. Regression test t_printf_length
covers the wide conversions, the cursor alignment, and the
truncating h modifiers.

Fixes #332.

+++ kernel/src/hardware/timer.c
@@ -560,12 +560,10 @@ void run_timer_softirq(void)
 /// @param data The data.
 static inline void debug_timeout(unsigned long data)
 {
-    // The seconds come back as a uint64_t and the logging printf has no
-    // 64-bit conversion, so narrow it here (#270).
     pr_debug(
         "The timer has been successfully deactivated: %d, ticks: %d, seconds: "
-        "%u\n",
-        data, timer_ticks, (uint32_t)timer_get_seconds());
+        "%llu\n",
+        data, timer_ticks, timer_get_seconds());
 }
 
 /// @brief Cancels a sleep timer for a task being woken up by signals.
